requirements.txt
numpy==1.23.5
scikit_learn==1.2.2
Pillow==10.0.0
pydicom==2.3.1
scipy==1.10.1
opencv_python==4.7.0.72
torch==2.0.1
ultralytics==8.0.111